Move getEnvironmentVariables() to AbstractCloudFoundryDeployer to ensure `use-spring-application-json` is consistently applied
Move stream specific properties to CloudFoundryAppDeployer

Move stream commandlineArgs to CloudFoundryAppDeployer
