+++ src/main/java/org/springframework/cloud/deployer/spi/cloudfoundry/CloudFoundryAppDeployer.java
@@ -18,7 +18,6 @@ package org.springframework.cloud.deployer.spi.cloudfoundry;
 
 import java.time.Duration;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -26,8 +25,6 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.cloudfoundry.doppler.LogMessage;
 import org.cloudfoundry.operations.CloudFoundryOperations;
 import org.cloudfoundry.operations.applications.ApplicationDetail;
@@ -65,7 +62,6 @@ import static org.springframework.cloud.deployer.spi.cloudfoundry.CloudFoundryDe
 import static org.springframework.cloud.deployer.spi.cloudfoundry.CloudFoundryDeploymentProperties.ROUTES_PROPERTY;
 import static org.springframework.cloud.deployer.spi.cloudfoundry.CloudFoundryDeploymentProperties.ROUTE_PATH_PROPERTY;
 import static org.springframework.cloud.deployer.spi.cloudfoundry.CloudFoundryDeploymentProperties.ROUTE_PROPERTY;
-import static org.springframework.cloud.deployer.spi.cloudfoundry.CloudFoundryDeploymentProperties.USE_SPRING_APPLICATION_JSON_KEY;
 
 
 /**
@@ -79,7 +75,7 @@ import static org.springframework.cloud.deployer.spi.cloudfoundry.CloudFoundryDe
  */
 public class CloudFoundryAppDeployer extends AbstractCloudFoundryDeployer implements MultiStateAppDeployer {
 
-	private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
+
 
 	private static final Logger logger = LoggerFactory.getLogger(CloudFoundryAppDeployer.class);
 
@@ -148,6 +144,31 @@ public class CloudFoundryAppDeployer extends AbstractCloudFoundryDeployer implem
 		} else {
 			return DeploymentState.undeployed;
 		}
+	}
+
+	@Override
+	protected Map<String, String> getEnvironmentVariables(String deploymentId, AppDeploymentRequest request) {
+		Map<String, String> envVariables = super.getEnvironmentVariables(deploymentId, request);
+		envVariables.putAll(getCommandLineArguments(request));
+		String group = request.getDeploymentProperties().get(AppDeployer.GROUP_PROPERTY_KEY);
+		if (StringUtils.hasText(group)) {
+			envVariables.put("SPRING_CLOUD_APPLICATION_GROUP", group);
+		}
+		envVariables.put("SPRING_CLOUD_APPLICATION_GUID", "${vcap.application.name}:${vcap.application.instance_index}");
+		envVariables.put("SPRING_APPLICATION_INDEX", "${vcap.application.instance_index}");
+		return envVariables;
+	}
+
+	private Map<String, String> getCommandLineArguments(AppDeploymentRequest request) {
+		if (request.getCommandlineArguments().isEmpty()) {
+			return Collections.emptyMap();
+		}
+
+		String argumentsAsString = request.getCommandlineArguments().stream()
+				.collect(Collectors.joining(" "));
+		String yaml = new Yaml().dump(Collections.singletonMap("arguments", argumentsAsString));
+
+		return Collections.singletonMap("JBP_CONFIG_JAVA_MAIN", yaml);
 	}
 
 	@Override
@@ -249,58 +270,6 @@ public class CloudFoundryAppDeployer extends AbstractCloudFoundryDeployer implem
 			.orElse(this.deploymentProperties.getDomain());
 	}
 
-	private Map<String, String> getApplicationProperties(String deploymentId, AppDeploymentRequest request) {
-		Map<String, String> applicationProperties = getSanitizedApplicationProperties(deploymentId, request);
-
-		if (!useSpringApplicationJson(request)) {
-			return applicationProperties;
-		}
-
-		try {
-			return Collections.singletonMap("SPRING_APPLICATION_JSON", OBJECT_MAPPER.writeValueAsString(applicationProperties));
-		} catch (JsonProcessingException e) {
-			throw new RuntimeException(e);
-		}
-	}
-
-	private Map<String, String> getCommandLineArguments(AppDeploymentRequest request) {
-		if (request.getCommandlineArguments().isEmpty()) {
-			return Collections.emptyMap();
-		}
-
-		String argumentsAsString = request.getCommandlineArguments().stream()
-			.collect(Collectors.joining(" "));
-		String yaml = new Yaml().dump(Collections.singletonMap("arguments", argumentsAsString));
-
-		return Collections.singletonMap("JBP_CONFIG_JAVA_MAIN", yaml);
-	}
-
-	private Map<String, String> getEnvironmentVariables(String deploymentId, AppDeploymentRequest request) {
-		Map<String, String> envVariables = new HashMap<>();
-		envVariables.putAll(getApplicationProperties(deploymentId, request));
-		envVariables.putAll(getCommandLineArguments(request));
-		String javaOpts = javaOpts(request);
-		if (StringUtils.hasText(javaOpts)) {
-			envVariables.put("JAVA_OPTS", javaOpts(request));
-		}
-		String group = request.getDeploymentProperties().get(AppDeployer.GROUP_PROPERTY_KEY);
-		if (StringUtils.hasText(group)) {
-			envVariables.put("SPRING_CLOUD_APPLICATION_GROUP", group);
-		}
-		envVariables.put("SPRING_CLOUD_APPLICATION_GUID", "${vcap.application.name}:${vcap.application.instance_index}");
-		envVariables.put("SPRING_APPLICATION_INDEX", "${vcap.application.instance_index}");
-		return envVariables;
-	}
-
-	private Map<String, String> getSanitizedApplicationProperties(String deploymentId, AppDeploymentRequest request) {
-		Map<String, String> applicationProperties = new HashMap<>(request.getDefinition().getProperties());
-
-		// Remove server.port as CF assigns a port for us, and we don't want to override that
-		Optional.ofNullable(applicationProperties.remove("server.port"))
-			.ifPresent(port -> logger.warn("Ignoring 'server.port={}' for app {}, as Cloud Foundry will assign a local dynamic port. Route to the app will use port 80.", port, deploymentId));
-
-		return applicationProperties;
-	}
 
 	private Mono<AppStatus> getStatus(String deploymentId) {
 		return requestGetApplication(deploymentId)
@@ -482,10 +451,5 @@ public class CloudFoundryAppDeployer extends AbstractCloudFoundryDeployer implem
 			.orElse(null);
 	}
 
-	private boolean useSpringApplicationJson(AppDeploymentRequest request) {
-		return Optional.ofNullable(request.getDeploymentProperties().get(USE_SPRING_APPLICATION_JSON_KEY))
-			.map(Boolean::valueOf)
-			.orElse(this.deploymentProperties.isUseSpringApplicationJson());
-	}
 
 }
